invextrack/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-storages>=1.13.0
django-notifications-hq>=1.7.1
django-debug-toolbar>=3.9.0
django-filter>=23.2
django-allauth>=0.54.0
django-compressor>=4.4